+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved INFO logging on auto-diskful and delete resources
+- Try to keep tiebreaker on auto-diskful delete excessive resources
 
 ## [1.32.0-rc.3] - 2025-08-26