+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- SQL: SET SCHEMA statement triggered an early transaction start before any data was manipulated
 - ScopeRunner: start DB transaction after scope locks are acquired
 
 ## [1.32.0-rc.3] - 2025-08-26