+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.32.0] - 2025-09-09
+
 ### Changed
 
 - Improved INFO logging on auto-diskful and delete resources